yolo_ws_api/requirements.txt
ultralytics>=8.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
python-multipart>=0.0.6